feat(ehr): more fixes
Ref: #1040
