+++ packages/api/src/command/mapping/cx.ts
@@ -1,8 +1,17 @@
 import { uuidv7 } from "@metriport/core/util/uuid-v7";
-import { NotFoundError } from "@metriport/shared";
-import { CxMapping, CxMappingPerSource, CxMappingSource } from "../../domain/cx-mapping";
+import { NotFoundError, MetriportError } from "@metriport/shared";
+import {
+  CxMapping,
+  CxMappingPerSource,
+  CxMappingSource,
+  isCxMappingSource,
+} from "../../domain/cx-mapping";
+import {
+  isEhrClientJwtTokenSource,
+  isEhrDashJwtTokenSource,
+  isEhrWebhookJwtTokenSource,
+} from "../../external/ehr/shared";
 import { CxMappingModel } from "../../models/cx-mapping";
-
 export type CxMappingParams = CxMappingPerSource;
 
 export type CxMappingLookUpParams = Omit<CxMappingParams, "cxId" | "secondaryMappings">;
@@ -111,4 +120,23 @@ export async function setExternalIdOnCxMapping({
 export async function deleteCxMapping({ cxId, id }: CxMappingLookupByIdParams): Promise<void> {
   const existing = await getCxMappingModelByIdOrFail({ cxId, id });
   await existing.destroy();
+}
+
+export function getCxMappingSourceFromJwtTokenSource(source: string): CxMappingSource {
+  const additionalDetails = { source };
+  if (isEhrDashJwtTokenSource(source)) {
+    if (isCxMappingSource(source)) return source;
+    throw new MetriportError("Invalid dash source", undefined, additionalDetails);
+  }
+  if (isEhrClientJwtTokenSource(source)) {
+    const sourceWithoutClient = source.replace("-client", "");
+    if (isCxMappingSource(sourceWithoutClient)) return sourceWithoutClient;
+    throw new MetriportError("Invalid client source", undefined, additionalDetails);
+  }
+  if (isEhrWebhookJwtTokenSource(source)) {
+    const sourceWithoutWebhook = source.replace("-webhook", "");
+    if (isCxMappingSource(sourceWithoutWebhook)) return sourceWithoutWebhook;
+    throw new MetriportError("Invalid webhook source", undefined, additionalDetails);
+  }
+  throw new MetriportError("Invalid source", undefined, additionalDetails);
 }
