+++ packages/api/src/routes/ehr/athenahealth/patient.ts
@@ -1,10 +1,10 @@
+import { Request, Response } from "express";
 import Router from "express-promise-router";
 import httpStatus from "http-status";
-import { Request, Response } from "express";
 import { syncAthenaPatientIntoMetriport } from "../../../external/ehr/athenahealth/command/sync-patient";
+import { handleParams } from "../../helpers/handle-params";
 import { requestLogger } from "../../helpers/request-logger";
 import { asyncHandler, getCxIdOrFail, getFrom, getFromQueryOrFail } from "../../util";
-import { handleParams } from "../../helpers/handle-params";
 
 const router = Router();
